+++ packages/cli/src/lib/autorun/autorun-command.ts
@@ -6,8 +6,8 @@ import {
   collectAndPersistReports,
   upload,
 } from '@code-pushup/core';
-import { CLI_NAME } from '../cli';
-import { onlyPluginsOption } from '../implementation/only-plugins-options';
+import { CLI_NAME } from '../constants';
+import { yargsOnlyPluginsOptionsDefinition } from '../implementation/only-plugins.options';
 
 type AutorunOptions = CollectOptions & UploadOptions;
 
@@ -16,9 +16,7 @@ export function yargsAutorunCommandObject() {
   return {
     command,
     describe: 'Shortcut for running collect followed by upload',
-    builder: {
-      onlyPlugins: onlyPluginsOption,
-    },
+    builder: yargsOnlyPluginsOptionsDefinition(),
     handler: async <T>(args: ArgumentsCamelCase<T>) => {
       console.info(chalk.bold(CLI_NAME));
       console.info(chalk.gray(`Run ${command}...`));
@@ -36,10 +34,11 @@ export function yargsAutorunCommandObject() {
       };
 
       await collectAndPersistReports(optionsWithFormat);
-      if (!options.upload) {
-        console.warn('Upload skipped because configuration is not set.'); // @TODO log verbose
-      } else {
+
+      if (options.upload) {
         await upload(options);
+      } else {
+        console.warn('Upload skipped because configuration is not set.'); // @TODO log verbose
       }
     },
   } satisfies CommandModule;