+++ packages/cli/src/lib/yargs-cli.integration.test.ts
@@ -1,9 +1,9 @@
 import { describe, expect, it } from 'vitest';
 import { CoreConfig } from '@code-pushup/models';
-import {
-  CoreConfigCliOptions,
-  GeneralCliOptions,
-} from './implementation/model';
+import { CoreConfigCliOptions } from './implementation/core-config.model';
+import { GeneralCliOptions } from './implementation/global.model';
+import { OnlyPluginsOptions } from './implementation/only-plugins.model';
+import { yargsOnlyPluginsOptionsDefinition } from './implementation/only-plugins.options';
 import { options } from './options';
 import { yargsCli } from './yargs-cli';
 
@@ -21,6 +21,14 @@ describe('yargsCli', () => {
       options,
     }).parseAsync();
     expect(parsedArgv.config).toBe('code-pushup.config.js');
+  });
+
+  it('should parse an empty array as a default onlyPlugins option', async () => {
+    const parsedArgv = await yargsCli<GeneralCliOptions & OnlyPluginsOptions>(
+      [],
+      { options: { ...options, ...yargsOnlyPluginsOptionsDefinition() } },
+    ).parseAsync();
+    expect(parsedArgv.onlyPlugins).toEqual([]);
   });
 
   it('should parse a single boolean negated argument', async () => {
@@ -43,7 +51,16 @@ describe('yargsCli', () => {
       ['--persist.format=md', '--persist.format=json'],
       { options },
     ).parseAsync();
-    expect(parsedArgv.persist.format).toEqual(['md', 'json']);
+    expect(parsedArgv.persist?.format).toEqual(['md', 'json']);
+  });
+
+  it('should throw for an invalid persist format', () => {
+    expect(() =>
+      yargsCli<CoreConfig>(['--persist.format=md', '--persist.format=stdout'], {
+        options,
+        noExitProcess: true,
+      }).parse(),
+    ).toThrow('Invalid persist.format option');
   });
 
   it('should parse global options correctly', async () => {
@@ -64,7 +81,9 @@ describe('yargsCli', () => {
   });
 
   it('should handle global options and middleware argument overrides correctly', async () => {
-    const parsedArgv = await yargsCli<GeneralCliOptions & CoreConfigCliOptions>(
+    const parsedArgv = await yargsCli<
+      GeneralCliOptions & CoreConfigCliOptions & OnlyPluginsOptions
+    >(
       [
         '--verbose',
         '--persist.format=md',
@@ -74,8 +93,10 @@ describe('yargsCli', () => {
         '--upload.project=code-push-down',
         '--upload.server=https://code-pushdown.com/api',
         '--upload.apiKey=some-api-key',
+        '--onlyPlugins=lighthouse',
+        '--onlyPlugins=eslint',
       ],
-      { options },
+      { options: { ...options, ...yargsOnlyPluginsOptionsDefinition() } },
     ).parseAsync();
     expect(parsedArgv).toEqual(
       expect.objectContaining({
@@ -95,6 +116,7 @@ describe('yargsCli', () => {
           server: 'https://code-pushdown.com/api',
           apiKey: 'some-api-key',
         }),
+        onlyPlugins: ['lighthouse', 'eslint'],
       }),
     );
   });