+++ packages/cli/src/lib/yargs-cli.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines-per-function */
 import chalk from 'chalk';
 import yargs, {
   Argv,
@@ -6,8 +7,9 @@ import yargs, {
   Options,
   ParserConfigurationOptions,
 } from 'yargs';
+import { PersistConfig, formatSchema } from '@code-pushup/models';
 import { MAX_SCREEN_WIDTH } from '@code-pushup/models';
-import { logErrorBeforeThrow } from './implementation/utils';
+import { logErrorBeforeThrow } from './implementation/global.utils';
 
 /**
  * returns configurable yargs CLI for code-pushup
@@ -32,10 +34,9 @@ export function yargsCli<T = unknown>(
   },
 ): Argv<T> {
   const { usageMessage, scriptName, noExitProcess } = cfg;
-  let { commands, options, middlewares } = cfg;
-  commands = Array.isArray(commands) ? commands : [];
-  middlewares = Array.isArray(middlewares) ? middlewares : [];
-  options = options || {};
+  const commands = cfg.commands ?? [];
+  const middlewares = cfg.middlewares ?? [];
+  const options = cfg.options ?? {};
   const cli = yargs(argv);
 
   // setup yargs
@@ -43,19 +44,20 @@ export function yargsCli<T = unknown>(
     .help()
     .version(false)
     .alias('h', 'help')
+    .check(args => {
+      const persist = args['persist'] as PersistConfig | undefined;
+      return persist == null || validatePersistFormat(persist);
+    })
     .parserConfiguration({
       'strip-dashed': true,
     } satisfies Partial<ParserConfigurationOptions>)
-    .array('persist.format')
-    .coerce('config', (config: string | string[]) => {
-      if (Array.isArray(config)) {
-        return config[config.length - 1];
-      }
-      return config;
-    })
+    .coerce('config', (config: string | string[]) =>
+      Array.isArray(config) ? config.at(-1) : config,
+    )
     .options(options)
     // take full width of the terminal `cli.terminalWidth()`
     .wrap(MAX_SCREEN_WIDTH);
+
   // usage message
   if (usageMessage) {
     cli.usage(chalk.bold(usageMessage));
@@ -78,9 +80,7 @@ export function yargsCli<T = unknown>(
   commands.forEach(commandObj => {
     cli.command({
       ...commandObj,
-      ...(commandObj.handler && {
-        handler: logErrorBeforeThrow(commandObj.handler),
-      }),
+      handler: logErrorBeforeThrow(commandObj.handler),
       ...(typeof commandObj.builder === 'function' && {
         builder: logErrorBeforeThrow(commandObj.builder),
       }),
@@ -96,4 +96,19 @@ export function yargsCli<T = unknown>(
 
   // return CLI object
   return cli as unknown as Argv<T>;
+}
+
+function validatePersistFormat(persist: PersistConfig) {
+  try {
+    if (persist?.format != null) {
+      persist.format.forEach(format => formatSchema.parse(format));
+    }
+    return true;
+  } catch {
+    throw new Error(
+      `Invalid persist.format option. Valid options are: ${Object.values(
+        formatSchema.Values,
+      ).join(', ')}`,
+    );
+  }
 }