+++ packages/models/src/lib/implementation/utils.unit.test.ts
@@ -7,7 +7,6 @@ import {
 } from './utils';
 
 describe('slugRegex', () => {
-  // test valid and array of strings against slugRegex with it blocks
   it.each([
     'hello',
     'hello-world',
@@ -15,11 +14,10 @@ describe('slugRegex', () => {
     '123',
     '123-456',
     '123-world-789',
-  ])(`should match valid %p`, slug => {
+  ])(`should match a valid slug %p`, slug => {
     expect(slug).toMatch(slugRegex);
   });
 
-  // test invalid and array of strings against slugRegex with it blocks
   it.each([
     '',
     ' ',
@@ -32,21 +30,19 @@ describe('slugRegex', () => {
     '123-',
     '-123',
     '123--456',
-  ])(`should not match invalid slugs %p`, invalidSlugs => {
-    expect(invalidSlugs).not.toMatch(slugRegex);
+  ])(`should not match an invalid slug %p`, invalidSlug => {
+    expect(invalidSlug).not.toMatch(slugRegex);
   });
 });
 
 describe('filenameRegex', () => {
-  // test valid and array of strings against filenameRegex with it blocks
   it.each(['report', 'report.mock', 'report-test.mock'])(
-    `should match valid %p`,
+    `should match a valid file name %p`,
     filename => {
       expect(filename).toMatch(filenameRegex);
     },
   );
 
-  // test invalid and array of strings against filenameRegex with it blocks
   it.each([
     '',
     ' ',
@@ -58,27 +54,38 @@ describe('filenameRegex', () => {
     'file<name',
     'file>name',
     'file|name',
-  ])(`should not match invalid file name %p`, invalidFilename => {
+  ])(`should not match an invalid file name %p`, invalidFilename => {
     expect(invalidFilename).not.toMatch(filenameRegex);
   });
 });
 
-describe('stringUnique', () => {
-  it('should return true for a list of unique strings', () => {
+describe('hasDuplicateStrings', () => {
+  it('should return false for a list of unique strings', () => {
     expect(hasDuplicateStrings(['a', 'b'])).toBe(false);
   });
-  it('should return a list of duplicated strings for a invalid list', () => {
+
+  it('should return a list of duplicates for a list with duplicates', () => {
     expect(hasDuplicateStrings(['a', 'b', 'a', 'c'])).toEqual(['a']);
   });
-  it('should return a false for a list with 1 item', () => {
+
+  it('should return false for a list with 1 item', () => {
     expect(hasDuplicateStrings(['a'])).toBe(false);
   });
 });
 
-describe('stringsExist', () => {
-  it('should return true for the strings exist in target array', () => {
+describe('hasMissingStrings', () => {
+  it('should return false for two identical arrays', () => {
     expect(hasMissingStrings(['a', 'b'], ['a', 'b'])).toBe(false);
   });
+
+  it('should return false for an array subset', () => {
+    expect(hasMissingStrings(['b'], ['a', 'b'])).toBe(false);
+  });
+
+  it('should return false for two empty arrays', () => {
+    expect(hasMissingStrings([], [])).toBe(false);
+  });
+
   it('should return a list of strings from source that are missing in target', () => {
     expect(hasMissingStrings(['a', 'b'], ['a', 'c'])).toEqual(['b']);
   });