+++ packages/models/src/lib/plugin-config.ts
@@ -1,4 +1,6 @@
 import { z } from 'zod';
+import { pluginAuditsSchema } from './audit';
+import { groupsSchema } from './group';
 import {
   materialIconSchema,
   metaSchema,
@@ -6,16 +8,9 @@ import {
   slugSchema,
 } from './implementation/schemas';
 import { errorItems, hasMissingStrings } from './implementation/utils';
-import { pluginAuditsSchema } from './plugin-config-audits';
-import { auditGroupsSchema } from './plugin-config-groups';
-import {
-  runnerConfigSchema,
-  runnerFunctionSchema,
-} from './plugin-config-runner';
+import { runnerConfigSchema, runnerFunctionSchema } from './runner-config';
 
-export const pluginMetaSchema = packageVersionSchema({
-  optional: true,
-})
+export const pluginMetaSchema = packageVersionSchema()
   .merge(
     metaSchema({
       titleDescription: 'Descriptive name',
@@ -26,7 +21,7 @@ export const pluginMetaSchema = packageVersionSchema({
   )
   .merge(
     z.object({
-      slug: slugSchema('References plugin. ID (unique within core config)'),
+      slug: slugSchema('Unique plugin slug within core config'),
       icon: materialIconSchema,
     }),
   );
@@ -35,7 +30,7 @@ export type PluginMeta = z.infer<typeof pluginMetaSchema>;
 export const pluginDataSchema = z.object({
   runner: z.union([runnerConfigSchema, runnerFunctionSchema]),
   audits: pluginAuditsSchema,
-  groups: auditGroupsSchema,
+  groups: groupsSchema,
 });
 type PluginData = z.infer<typeof pluginDataSchema>;
 
@@ -54,19 +49,16 @@ export type PluginConfig = z.infer<typeof pluginConfigSchema>;
 // helper for validator: every listed group ref points to an audit within the plugin
 function missingRefsFromGroupsErrorMsg(pluginCfg: PluginData) {
   const missingRefs = getMissingRefsFromGroups(pluginCfg);
-  return `In the groups, the following audit ref's needs to point to a audit in this plugin config: ${errorItems(
+  return `The following group references need to point to an existing audit in this plugin config: ${errorItems(
     missingRefs,
   )}`;
 }
 
 function getMissingRefsFromGroups(pluginCfg: PluginData) {
-  if (pluginCfg?.groups?.length && pluginCfg?.audits?.length) {
-    const groups = pluginCfg?.groups || [];
-    const audits = pluginCfg?.audits || [];
-    return hasMissingStrings(
-      groups.flatMap(({ refs: audits }) => audits.map(({ slug: ref }) => ref)),
-      audits.map(({ slug }) => slug),
-    );
-  }
-  return false;
+  return hasMissingStrings(
+    pluginCfg.groups?.flatMap(({ refs: audits }) =>
+      audits.map(({ slug: ref }) => ref),
+    ) ?? [],
+    pluginCfg.audits.map(({ slug }) => slug),
+  );
 }