+++ packages/models/src/lib/upload-config.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { urlSchema } from './implementation/schemas';
+import { slugSchema, urlSchema } from './implementation/schemas';
 
 export const uploadConfigSchema = z.object({
   server: urlSchema('URL of deployed portal API'),
@@ -7,12 +7,8 @@ export const uploadConfigSchema = z.object({
     description:
       'API key with write access to portal (use `process.env` for security)',
   }),
-  organization: z.string({
-    description: 'Organization in code versioning system',
-  }),
-  project: z.string({
-    description: 'Project in code versioning system',
-  }),
+  organization: slugSchema('Organization slug from Code PushUp portal'),
+  project: slugSchema('Project slug from Code PushUp portal'),
 });
 
 export type UploadConfig = z.infer<typeof uploadConfigSchema>;