+++ packages/utils/src/lib/transform.unit.test.ts
@@ -8,6 +8,7 @@ import {
   objectToEntries,
   objectToKeys,
   toArray,
+  toUnixPath,
 } from './transform';
 
 describe('toArray', () => {
@@ -162,5 +163,27 @@ describe('objectToCliArgs', () => {
   it('should throw error for unsupported type', () => {
     const params = { unsupported: undefined as any };
     expect(() => objectToCliArgs(params)).toThrow('Unsupported type');
+  });
+});
+
+describe('toUnixPath', () => {
+  it.each([
+    ['main.ts', 'main.ts'],
+    ['src/main.ts', 'src/main.ts'],
+    ['../../relative/unix/path/index.ts', '../../relative/unix/path/index.ts'],
+    [
+      '..\\..\\relative\\windows\\path\\index.ts',
+      '../../relative/windows/path/index.ts',
+    ],
+  ])('should transform "%s" to valid slug "%s"', (path, unixPath) => {
+    expect(toUnixPath(path)).toBe(unixPath);
+  });
+
+  it('should transform absolute Windows path to relative UNIX path', () => {
+    expect(
+      toUnixPath(`${process.cwd()}\\windows\\path\\config.ts`, {
+        toRelative: true,
+      }),
+    ).toBe('windows/path/config.ts');
   });
 });