+++ testing-utils/src/lib/testing-utils.unit.test.ts
@@ -1,7 +1,0 @@
-import { testingUtils } from './testing-utils';
-
-describe('testingUtils', () => {
-  it('should work', () => {
-    expect(testingUtils()).toBe('testing-utils');
-  });
-});