+++ testing-utils/src/lib/utils/dynamic-mocks/persist-config.mock.ts
@@ -1,0 +1,18 @@
+import {
+  PERSIST_FILENAME,
+  PERSIST_FORMAT,
+  PERSIST_OUTPUT_DIR,
+  PersistConfig,
+  persistConfigSchema,
+} from '@code-pushup/models';
+
+export function persistConfigMock(
+  opt?: Partial<PersistConfig>,
+): Required<PersistConfig> {
+  return persistConfigSchema.parse({
+    outputDir: PERSIST_OUTPUT_DIR,
+    filename: PERSIST_FILENAME,
+    format: PERSIST_FORMAT,
+    ...opt,
+  }) as Required<PersistConfig>;
+}