+++ CHANGELOG.md
@@ -2,6 +2,62 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.8.17](https://github.com/code-pushup/cli/compare/v0.8.16...v0.8.17) (2024-01-15)
+
+## [0.8.16](https://github.com/code-pushup/cli/compare/v0.8.15...v0.8.16) (2024-01-15)
+
+## [0.8.15](https://github.com/code-pushup/cli/compare/v0.8.14...v0.8.15) (2024-01-15)
+
+## [0.8.14](https://github.com/code-pushup/cli/compare/v0.8.13...v0.8.14) (2024-01-12)
+
+## [0.8.13](https://github.com/code-pushup/cli/compare/v0.8.12...v0.8.13) (2024-01-12)
+
+## [0.8.12](https://github.com/code-pushup/cli/compare/v0.8.11...v0.8.12) (2024-01-12)
+
+## [0.8.11](https://github.com/code-pushup/cli/compare/v0.8.10...v0.8.11) (2024-01-12)
+
+### Bug Fixes
+
+- **models:** fix use case for a group reference but empty audits, clean up tests ([d311b95](https://github.com/code-pushup/cli/commit/d311b956cf6d6c9532de30bfd4080bfeb37ab1c3))
+
+## [0.8.10](https://github.com/code-pushup/cli/compare/v0.8.9...v0.8.10) (2024-01-12)
+
+## [0.8.9](https://github.com/code-pushup/cli/compare/v0.8.8...v0.8.9) (2024-01-11)
+
+## [0.8.8](https://github.com/code-pushup/cli/compare/v0.8.7...v0.8.8) (2024-01-11)
+
+### Bug Fixes
+
+- **plugin-eslint:** truncate long issue messages ([c39d0cc](https://github.com/code-pushup/cli/commit/c39d0ccefbbc9f44a7ecb4d90346bca5b1553766))
+
+## [0.8.7](https://github.com/code-pushup/cli/compare/v0.8.6...v0.8.7) (2024-01-10)
+
+## [0.8.6](https://github.com/code-pushup/cli/compare/v0.8.5...v0.8.6) (2024-01-10)
+
+## [0.8.5](https://github.com/code-pushup/cli/compare/v0.8.4...v0.8.5) (2024-01-10)
+
+## [0.8.4](https://github.com/code-pushup/cli/compare/v0.8.3...v0.8.4) (2024-01-10)
+
+## [0.8.3](https://github.com/code-pushup/cli/compare/v0.8.2...v0.8.3) (2024-01-10)
+
+## [0.8.2](https://github.com/code-pushup/cli/compare/v0.8.1...v0.8.2) (2024-01-09)
+
+## [0.8.1](https://github.com/code-pushup/cli/compare/v0.8.0...v0.8.1) (2024-01-08)
+
+# [0.8.0](https://github.com/code-pushup/cli/compare/v0.7.0...v0.8.0) (2024-01-08)
+
+### Features
+
+- **core:** add grouping of promise results by status ([5f50e8e](https://github.com/code-pushup/cli/commit/5f50e8ecafa6742c597604bf3ffdcd0361f8c128)), closes [#287](https://github.com/code-pushup/cli/issues/287) [#287](https://github.com/code-pushup/cli/issues/287)
+
+# [0.7.0](https://github.com/code-pushup/cli/compare/v0.6.6...v0.7.0) (2024-01-08)
+
+### Features
+
+- **core:** improve upload error messages, extend icon set ([2251ebd](https://github.com/code-pushup/cli/commit/2251ebd13763facb562a9011cf034ff2e9b7ef37))
+
+## [0.6.6](https://github.com/code-pushup/cli/compare/v0.6.5...v0.6.6) (2024-01-05)
+
 ## [0.6.5](https://github.com/code-pushup/cli/compare/v0.6.4...v0.6.5) (2024-01-04)
 
 ## [0.6.4](https://github.com/code-pushup/cli/compare/v0.6.3...v0.6.4) (2024-01-04)